backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
reportlab>=4.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
aiofiles>=23.1.0
